오목

+++ OMOK/OMOK/OMOK.Android/Dependency.cs
@@ -20,7 +20,7 @@ namespace ToastMessage.Droid
     {
         public void Show(string message)
         {
-            var toast = Android.Widget.Toast.MakeText(Android.App.Application.Context, message, ToastLength.Long);
+            var toast = Android.Widget.Toast.MakeText(Android.App.Application.Context, message, ToastLength.Short);
             toast.SetGravity(GravityFlags.Top | GravityFlags.Center, 0, 5);
             toast.Show();
         }
